Created FIles
20/11/2025
Created the basic files that are needed for both Part A and Part B. Added input boxes, they do not currently work due to the lack of JavaScript. Also added some styling found in style.css, mostly black background and white text for now, may change later. Finally, attached some of the theory that has been completed. Moving forward, the plan is to add function to the HTML file with some JavaScript.