Basic User Authentication
23/11/2025
Added basic user authentication, using the phrase 'password' as the password and 'admin' as the username for testing purposes. Note that Heading 1 is used as a response for when login is successful. Functional authentication will be added later, as of now this is just a check to see that everything works as intended.